Separate Table class from initialization code
