+++ CHANGELOG.md
@@ -9,6 +9,7 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 ### Changed
 - **cf-forms:** [PATCH] Changes label helper text to Avenir Regular.
 - **cf-forms:** [PATCH] Lets autoprefixer handle placeholder pseudoselector.
+- **cf-tables:** [PATCH] Separate Table class from initialization code.
 
 ### Removed
 -
