+++ src/cf-tables/src/cf-tables.js
@@ -1,27 +1,9 @@
 /* ==========================================================================
-   TableOrganism
-
+   Table initialization code.
    ========================================================================== */
 
 'use strict';
 
-var config = require( 'atomic-component/src/utilities/config' );
-var Organism = require( 'atomic-component/src/components/Organism' );
-var TableSortable = require( './cf-table-sortable' );
-var TableRowLinks = require( './cf-table-row-links' );
+const Table = require( './Table' );
 
-var TableOrganism = Organism.extend( {
-
-  ui: {
-    base: '.o-table'
-  },
-
-  modifiers: [ TableSortable, TableRowLinks ]
-
-} );
-
-TableOrganism.constants.DIRECTIONS = config.DIRECTIONS;
-
-TableOrganism.init();
-
-module.exports = TableOrganism;
+Table.init();
